# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminThreadResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_preview": {
            "title": "Last Preview",
            "type": "string"
          },
          "message_count": {
            "title": "Message Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_id",
          "agent_name",
          "user_id",
          "username",
          "title",
          "message_count",
          "last_preview",
          "created_at",
          "updated_at"
        ],
        "title": "AdminThreadResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "properties": {
          "text": {
            "maxLength": 32000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "description": "Mensaje del transcript (D16). Solo `user`/`assistant`.",
        "properties": {
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "ThreadCreate": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "title": {
            "default": "Nueva conversación",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "ThreadCreate",
        "type": "object"
      },
      "ThreadResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_consolidated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Consolidated At"
          },
          "last_preview": {
            "title": "Last Preview",
            "type": "string"
          },
          "message_count": {
            "title": "Message Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "agent_id",
          "user_id",
          "title",
          "message_count",
          "last_preview",
          "created_at",
          "updated_at",
          "last_consolidated_at"
        ],
        "title": "ThreadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "theyrethink-ai",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/threads": {
      "get": {
        "operationId": "list_all_threads_api_v1_admin_threads_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminThreadResponse"
                  },
                  "title": "Response List All Threads Api V1 Admin Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All Threads",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/threads/{thread_id}/messages": {
      "get": {
        "description": "Transcript de cualquier hilo, leído del checkpointer (D16/D18).",
        "operationId": "list_thread_messages_api_v1_admin_threads__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Thread Messages Api V1 Admin Threads  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Thread Messages",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/threads": {
      "get": {
        "operationId": "list_threads_api_v1_threads_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreadResponse"
                  },
                  "title": "Response List Threads Api V1 Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Threads",
        "tags": [
          "threads"
        ]
      },
      "post": {
        "operationId": "create_thread_api_v1_threads_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Thread",
        "tags": [
          "threads"
        ]
      }
    },
...
    "/api/v1/threads/{thread_id}/messages": {
      "get": {
        "description": "Transcript del hilo, leido del checkpointer via `agent/service.py` (D16).",
        "operationId": "list_messages_api_v1_threads__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Api V1 Threads  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "operationId": "send_message_api_v1_threads__thread_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "theyrethink_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theyrethink Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}